tests: cover authorization in REST API
