+++ tests/unit/serve/routes/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for slurm-quota serve HTTP routes."""
